Added Semester Endpoint as per issue #69. Endpoint should be able to insert semesters into the db.

+++ WhoteachswhatJPA/src/main/java/com/seneca/repository/SemesterDao.java
@@ -8,15 +8,16 @@ public interface SemesterDao
 {
    public List<Semester> getAll();
 
-   public Semester getById(Long id);
+	public Semester getById(Integer id);
 
    public void update(Semester entity);
 
    public void create(Semester entity);
 
-   public void delete(Long id);
+	public void delete(Integer id);
 
-   public List<Semester> search(Semester search, Long first, Integer maxItems);
+	public List<Semester> search(Semester search, Integer first,
+			Integer maxItems);
 
    public Long getCount(Semester search);
 }
